feat/Voter for User

+++ high-riders/src/Controller/Api/v1/UserController.php
@@ -131,11 +131,11 @@ class UserController extends AbstractController
      */
     public function update(int $id, UserRepository $userRepository, User $user, Request $request, SerializerInterface $serialiser): Response
     {
-        // TODO: vérifier les voters
-        // $this->denyAccessUnlessGranted('edit', $user, "Vous n'avez pas accés à cette page' !");
-
         // A user is retrieved according to its id
         $user = $userRepository->find($id);
+       
+        $this->denyAccessUnlessGranted('USER_EDIT', $user, "Vous n'avez pas accés à cette page' !");
+
         
         // If the user does not exist, we return a 404 error
         if (!$user) {
@@ -177,11 +177,11 @@ class UserController extends AbstractController
     public function delete(User $user): Response
     {
 
-        // if ($this->denyAccessUnlessGranted('delete', $user,"Vous n'avez pas accés à cette page' !")) {
+         if ($this->denyAccessUnlessGranted('USER_DELETE', $user,"Vous n'avez pas accés à cette page' !")) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($user);
             $entityManager->flush();
-        // }
+        }
 
         return $this->json(['l\'utilisateur à été suprimer'], 204);
     }
